Improve typing for PGNTurnList
This removes the generic type from PGNTurnList, as it's pretty hard to
represent the type properly in this way, considering it's recursive
nature. Instead, the type is now hard-coded as a sequence of moves or
itself, which removes the recursive type issues.

That said, this also isn't ideal, as it no longer allows for specifying
that the type only holds moves (no variations), which was useful for
flattened lines. That said, it is still a better solution, so even
through this downside, it's worth it.